refactor(semantic): move expl3 statement segmentation out of the formatter
The segmentation is analysis, not layout: "this head plus these
arguments form one call" is semantics layered on the syntax tree,
the same genre as `semantic::define`'s definition scan. It now
lives beside the slot model in `semantic::expl3`, and the
formatter consumes the statement map the way it consumes
signatures — only the gap policy (hard vs breakable, sticky vs
greedy, hangs) stays in `lower_expl_code`.

The shared shape predicates `is_collapsible_trivia` and
`is_param_digit` move from `formatter::core` to `crate::syntax`.

Behavior-neutral: full suite green and the `--checks all` gate
sets re-verified byte-identical for all three corpora. This also
positions the model as the differential oracle for the recorded
decision #8 candidate (arity-directed attachment, TODO.md).
